Urchin: - Randomised positions in bottom quarter - Added and stored in array
Other:
- Shortened initpos calculation (from ` X / 2 *-1' to ' X / -2')
- Changed OrangeFish speed datatype to float (to allow dividing xSpeed to .5's)

+++ FishORama/OrangeFish.cs
@@ -30,8 +30,8 @@ namespace FishORama
         // *** ADD YOUR CLASS VARIABLES HERE *** 
         Random rand;
 
-        int xSpeed;
-        int ySpeed;
+        float xSpeed;
+        float ySpeed;
         
         int assetHeight = 86;
         int assetWidth = 128;
